requirements.txt
bottle
gunicorn
weather-api
python-launch-library
